refactor: io 파일 이름 view로 수정
