+++ src/view/inputHandler.js
@@ -1,6 +1,6 @@
 import { INFO_MESSAGE } from "../config/constants.js";
-import readLineAsync from "./input.js";
 import { validateAttempt } from "../utils/validate.js";
+import readLineAsync from "./input.js";
 
 export const getCarNames = async () => {
   return (await readLineAsync(INFO_MESSAGE.CAR_NAME_INPUT)).trim();
