+++ src/view/outputHandler.js
